# bpo-47177: Replace `f_lasti` with `prev_instr`