--- a/Include/internal/pycore_frame.h
+++ b/Include/internal/pycore_frame.h
@@ -56,7 +56,7 @@ typedef struct _PyInterpreterFrame {
 
 // NOTE: This is not necessarily the index of the last instruction started in
 // the given frame. Rather, it is the index *prior to* the *next* instruction.
-// For example, it may the offset of an inline CACHE entry, an instruction we
+// For example, it may be the index of an inline CACHE entry, an instruction we
 // just jumped over, or (in the case of a newly-created frame) a value of -1.
 // It is here for mostly historical reasons.
 #define _PyInterpreterFrame_LASTI(IF) \