--- a/Objects/genobject.c
+++ b/Objects/genobject.c
@@ -1336,8 +1336,7 @@ compute_cr_origin(int origin_depth, _PyInterpreterFrame *current_frame)
     frame = current_frame;
     for (int i = 0; i < frame_count; ++i) {
         PyCodeObject *code = frame->f_code;
-        int addr = _PyInterpreterFrame_LASTI(frame) * sizeof(_Py_CODEUNIT);
-        int line = PyCode_Addr2Line(frame->f_code, addr);
+        int line = _PyInterpreterFrame_GetLine(frame);
         PyObject *frameinfo = Py_BuildValue("OiO", code->co_filename, line,
                                             code->co_name);
         if (!frameinfo) {