--- a/Python/frame.c
+++ b/Python/frame.c
@@ -127,3 +127,10 @@ _PyFrame_Push(PyThreadState *tstate, PyFunctionObject *func)
     _PyFrame_InitializeSpecials(new_frame, func, NULL, code->co_nlocalsplus);
     return new_frame;
 }
+
+int
+_PyInterpreterFrame_GetLine(_PyInterpreterFrame *frame)
+{
+    int addr = _PyInterpreterFrame_LASTI(frame) * sizeof(_Py_CODEUNIT);
+    return PyCode_Addr2Line(frame->f_code, addr);
+}