A nice 2% improvement:

<details>

```
Slower (2):
- telco: 38.6 ms +- 2.7 ms -> 40.9 ms +- 2.6 ms: 1.06x slower
- regex_dna: 1.33 sec +- 0.01 sec -> 1.34 sec +- 0.01 sec: 1.01x slower

Faster (32):
- logging_silent: 640 ns +- 27 ns -> 582 ns +- 35 ns: 1.10x faster
- scimark_sor: 736 ms +- 16 ms -> 689 ms +- 13 ms: 1.07x faster
- pycparser: 7.52 sec +- 0.10 sec -> 7.12 sec +- 0.10 sec: 1.06x faster
- chameleon: 39.8 ms +- 2.7 ms -> 37.8 ms +- 2.4 ms: 1.05x faster
- html5lib: 417 ms +- 35 ms -> 397 ms +- 31 ms: 1.05x faster
- scimark_monte_carlo: 422 ms +- 13 ms -> 402 ms +- 11 ms: 1.05x faster
- pyflate: 2.65 sec +- 0.02 sec -> 2.53 sec +- 0.02 sec: 1.05x faster
- deltablue: 22.5 ms +- 1.3 ms -> 21.6 ms +- 1.4 ms: 1.05x faster
- unpickle_pure_python: 1.43 ms +- 0.08 ms -> 1.37 ms +- 0.07 ms: 1.04x faster
- spectral_norm: 628 ms +- 10 ms -> 604 ms +- 13 ms: 1.04x faster
- hexiom: 41.0 ms +- 3.1 ms -> 39.5 ms +- 2.4 ms: 1.04x faster
- raytrace: 1.84 sec +- 0.02 sec -> 1.78 sec +- 0.02 sec: 1.03x faster
- scimark_lu: 669 ms +- 20 ms -> 648 ms +- 20 ms: 1.03x faster
- xml_etree_generate: 474 ms +- 11 ms -> 459 ms +- 13 ms: 1.03x faster
- richards: 296 ms +- 12 ms -> 287 ms +- 14 ms: 1.03x faster
- nbody: 555 ms +- 17 ms -> 540 ms +- 15 ms: 1.03x faster
- pickle_dict: 170 us +- 8 us -> 165 us +- 7 us: 1.03x faster
- crypto_pyaes: 510 ms +- 12 ms -> 497 ms +- 13 ms: 1.03x faster
- scimark_fft: 1.99 sec +- 0.03 sec -> 1.95 sec +- 0.02 sec: 1.02x faster
- regex_compile: 819 ms +- 11 ms -> 800 ms +- 10 ms: 1.02x faster
- chaos: 423 ms +- 12 ms -> 414 ms +- 10 ms: 1.02x faster
- meteor_contest: 643 ms +- 13 ms -> 630 ms +- 12 ms: 1.02x faster
- pidigits: 1.18 sec +- 0.01 sec -> 1.16 sec +- 0.01 sec: 1.02x faster
- sympy_expand: 2.86 sec +- 0.02 sec -> 2.81 sec +- 0.02 sec: 1.02x faster
- nqueens: 498 ms +- 13 ms -> 490 ms +- 13 ms: 1.02x faster
- xml_etree_parse: 953 ms +- 20 ms -> 938 ms +- 16 ms: 1.02x faster
- 2to3: 1.60 sec +- 0.01 sec -> 1.58 sec +- 0.01 sec: 1.02x faster
- xml_etree_process: 331 ms +- 12 ms -> 326 ms +- 11 ms: 1.02x faster
- sympy_str: 1.71 sec +- 0.02 sec -> 1.69 sec +- 0.02 sec: 1.01x faster
- fannkuch: 2.42 sec +- 0.02 sec -> 2.39 sec +- 0.02 sec: 1.01x faster
- dulwich_log: 381 ms +- 11 ms -> 377 ms +- 11 ms: 1.01x faster
- xml_etree_iterparse: 624 ms +- 14 ms -> 618 ms +- 13 ms: 1.01x faster

Benchmark hidden because not significant (28): django_template, float, go, json, json_dumps, json_loads, logging_format, logging_simple, mako, pathlib, pickle, pickle_list, pickle_pure_python, python_startup, python_startup_no_site, regex_effbot, regex_v8, scimark_sparse_mat_mult, sqlalchemy_declarative, sqlalchemy_imperative, sqlite_synth, sympy_integrate, sympy_sum, thrift, tornado_http, unpack_sequence, unpickle, unpickle_list

Geometric mean: 1.02x faster
```

</details>

<!-- issue-number: [bpo-47177](https://bugs.python.org/issue47177) -->
https://bugs.python.org/issue47177
<!-- /issue-number -->