@markshannon 2022-03-31
[My performance test](https://gist.github.com/markshannon/62f3d9f37946085066f1fc38140c1895) shows the same 2% but with less variation, and no 6% slowdown for `telco`.