@markshannon 2022-04-06
Did you get the version with `prev_instr` ready for review/merging?