@markshannon 2022-04-06
Whatever is easier. I don't mind