@brandtbucher 2022-04-07
https://bugs.python.org/issue47250