@vstinner 2022-04-07
The coverage project uses the PyFrameObject.f_frame.f_lasti field and so is broken by this PR. I wrote https://github.com/nedbat/coveragepy/pull/1331 to read the Python attribute, but the coverage maintainer has worries about worse performance. Maybe adding a new PyFrame_GetLastInstr() getter function would help.