@brandtbucher 2022-04-07
The PR adds a `_PyInterpreterFrame_LASTI` macro. Does that help?