@markshannon 2022-04-08
https://github.com/python/cpython/pull/32413